chore: move DocumentSelector, DocumentUploadSection, StudentInfoCard to /job-apply
