--- src/components/StudentApplicant.tsx
@@ -1,19 +1,0 @@
-"use client";
-
-interface StudentApplicantProps {
-  studentProfile: string;
-  studentName: string;
-  studentEmail: string;
-  tags: string[];
-  date: string;
-}
-
-const StudentApplicantCard = ({studentProfile, studentName, studentEmail, 
-    tags, date}: StudentApplicantProps) => {
-        
-  return (
-  <div></div>
-  );
-};
-
-export default StudentApplicantCard;
